requirements.txt
Django==4.0.1
djangorestframework==3.13.1
dj-rest-auth==2.2.2
django-import-export==2.7.1
